+++ lib/asn1c/common/constr_SET_OF_aper.c
@@ -34,7 +34,9 @@ SET_OF_encode_aper(const asn_TYPE_descriptor_t *td,
     if(ct) {
         int not_in_root =
             (list->count < ct->lower_bound || list->count > ct->upper_bound);
-        ASN_DEBUG("lb %ld ub %ld %s", ct->lower_bound, ct->upper_bound,
+        ASN_DEBUG("lb %lld ub %lld %s",
+                  (long long int)ct->lower_bound,
+                  (long long int)ct->upper_bound,
                   ct->flags & APC_EXTENSIBLE ? "ext" : "fix");
         if(ct->flags & APC_EXTENSIBLE) {
             /* Declare whether size is in extension root */
@@ -130,8 +132,8 @@ SET_OF_decode_aper(const asn_codec_ctx_t *opt_codec_ctx,
     if(ct && ct->effective_bits >= 0) {
         /* X.691, #19.5: No length determinant */
         nelems = aper_get_nsnnwn(pd, ct->upper_bound - ct->lower_bound + 1);
-        ASN_DEBUG("Preparing to fetch %ld+%ld elements from %s",
-                  (long)nelems, ct->lower_bound, td->name);
+        ASN_DEBUG("Preparing to fetch %ld+%lld elements from %s",
+                  (long)nelems, (long long int)ct->lower_bound, td->name);
         if(nelems < 0)  ASN__DECODE_STARVED;
         nelems += ct->lower_bound;
     } else {
